feat: initial commit for SSR experimental branch

+++ src/app/core/services/storage.service.ts
@@ -1,11 +1,15 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { inject, Injectable, PLATFORM_ID } from '@angular/core';
 
 @Injectable({
   providedIn: 'root'
 })
 export class StorageService {
-
-  constructor() { }
+  private isBrowser: boolean;
+  private platformId = inject(PLATFORM_ID)
+  constructor() {
+    this.isBrowser = isPlatformBrowser(this.platformId);
+  }
 
   private handleError<T>(operation: string, defaultValue: T | null, error: any): T | null {
     console.error(`Failed to ${operation} in local storage: ${error}`);
@@ -13,6 +17,9 @@ export class StorageService {
   }
 
   public get<T>(key: string, defaultValue?: T): T | null {
+    if (!this.isBrowser) {
+      return defaultValue || null;
+    }
     try {
       const item = localStorage.getItem(key);
       if (item === null) {
@@ -25,6 +32,9 @@ export class StorageService {
   }
 
   public set<T>(key: string, value: T): boolean {
+    if (!this.isBrowser) {
+      return false;
+    }
     try {
       const stringValue = JSON.stringify(value);
       localStorage.setItem(key, stringValue);
@@ -36,6 +46,9 @@ export class StorageService {
   }
 
   public delete(key: string): void {
+    if (!this.isBrowser) {
+      return;
+    }
     try {
       localStorage.removeItem(key);
     } catch (error) {
@@ -44,6 +57,9 @@ export class StorageService {
   }
 
   public clearAll(): void {
+    if (!this.isBrowser) {
+      return;
+    }
     try {
       localStorage.clear();
     } catch (error) {
